\name{MCPMod-internal}
\alias{fitModel}
\alias{AIC.fitMCPMod}
\alias{predict.fitMCPMod}
\alias{coef.fitMCPMod}
\alias{pValues}
\alias{getTstat}
\alias{getInit}
\alias{getInitBeta}
\alias{getGrad}
\alias{getDoseEst}
\alias{getDose}
\alias{getDef}
\alias{modContr}
\alias{getOptCont}
\alias{getStdCont}
\alias{modelSelect}
\alias{getInitP}
\alias{getInitLRasymp}
\alias{recovNames}
\title{ Internal MCPMod functions }
\description{
  These functions are not meant to be called by the user.
}

\keyword{ internal }
