\name{plot.fullMod}
\alias{plot.fullMod}
\title{ Plot method for fullMod objects }
\description{
 Plot method for fullMod objects.
}
\usage{
\method{plot}{fullMod}(x, ...)
}
\arguments{
  \item{x}{ \code{fullMod} object }
  \item{\dots}{ Additional arguments for the \code{plotModels} function }
}
\seealso{ \code{\link{plotModels}}, \code{\link{fullMod}} }
\keyword{methods}
\keyword{hplot}
